separated post page and its view component logic
